pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=43.0.0", "wheel"]

[project]
authors = [{name = "Avnish Narayan", email = "avnish@anyscale.com"}]
dependencies = [
  "pydantic<2.5",
  "ray",
  "pytest>=6.0",
  "seaborn>=0.11",
  "awscli>=1.22",
  "typer>=0.4",
  "litellm>=0.1.738",
  "num2words",
  "transformers",
  "tqdm",
  "boto3",
  "google-cloud-aiplatform",
]
description = "A framework for load testing LLM APIs"
license = {text = "Apache-2.0"}
name = "LLMPerf"
requires-python = ">=3.8, <3.11"
version = "0.1.0"

[project.optional-dependencies]
azure_monitor = ["azure-monitor-opentelemetry-exporter==1.0.0b23"]
